Agregar eliminar cuenta, todavía no funciona

+++ prototipos/Prosefy/src/app/services/usuario.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from "@angular/common/http";
-import { Observable } from "rxjs";
+import { Observable, throwError } from "rxjs";
 
 export interface Usuario {
   username: string;
@@ -26,11 +26,19 @@ export class UsuarioService {
 
   getUsuarioPorId(usuarioId: string): Observable<Usuario> {
     const url = `${this.apiUrl}/${usuarioId}`;
-
     return this.http.get<Usuario>(url);
   }
 
   eliminarCuenta(idUsuario: string): Observable<any> {
+    // Verifica si el ID tiene el formato correcto
+    const isValidId = /^[0-9a-fA-F]{24}$/.test(idUsuario);
+
+    if (!isValidId) {
+      // Maneja el error o lanza una excepción según tu lógica
+      console.error('ID de usuario no válido');
+      return throwError('ID de usuario no válido');
+    }
+
     const url = `${this.apiUrl}/${idUsuario}`;
 
     return this.http.delete(url);
